refactor(game): take double for animation progress applying
We be comparing floating point number now.

+++ src/game/skin/skin_lr2_animation.h
@@ -8,7 +8,7 @@ namespace lunaticvibes
 
 // Formula matches LR2.
 // Keyword: gradient
-[[nodiscard]] inline constexpr double grad(int dst, int src, double t)
+[[nodiscard]] inline constexpr double grad(double dst, double src, double t)
 {
     return (src == dst) ? src : (dst * t + src * (1.0 - t));
 }
